+++ backend/src/main/java/com/woowacourse/levellog/authentication/presentation/OAuthController.java
@@ -1,7 +1,9 @@
 package com.woowacourse.levellog.authentication.presentation;
 
-import com.woowacourse.levellog.authentication.application.MemberLogin;
+import com.woowacourse.levellog.authentication.application.LoginService;
+import com.woowacourse.levellog.authentication.application.OAuthService;
 import com.woowacourse.levellog.authentication.dto.request.GithubCodeRequest;
+import com.woowacourse.levellog.authentication.dto.response.GithubProfileResponse;
 import com.woowacourse.levellog.authentication.dto.response.LoginResponse;
 import com.woowacourse.levellog.authentication.support.PublicAPI;
 import javax.validation.Valid;
@@ -18,10 +20,12 @@ import org.springframework.web.bind.annotation.RestController;
 @PublicAPI
 public class OAuthController {
 
-    private final MemberLogin memberLogin;
+    private final OAuthService oAuthService;
+    private final LoginService loginService;
 
     @PostMapping("/login")
     public ResponseEntity<LoginResponse> login(@RequestBody @Valid final GithubCodeRequest request) {
-        return ResponseEntity.ok(memberLogin.login(request));
+        final GithubProfileResponse githubProfile = oAuthService.login(request);
+        return ResponseEntity.ok(loginService.login(githubProfile));
     }
 }
