+++ backend/src/test/java/com/woowacourse/levellog/application/ServiceTest.java
@@ -4,7 +4,7 @@ import static com.woowacourse.levellog.fixture.TimeFixture.TEAM_START_TIME;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.woowacourse.levellog.admin.application.AdminService;
-import com.woowacourse.levellog.authentication.application.OAuthService;
+import com.woowacourse.levellog.authentication.application.LoginService;
 import com.woowacourse.levellog.authentication.support.JwtTokenProvider;
 import com.woowacourse.levellog.common.domain.BaseEntity;
 import com.woowacourse.levellog.common.dto.LoginStatus;
@@ -30,7 +30,6 @@ import com.woowacourse.levellog.prequestion.domain.PreQuestion;
 import com.woowacourse.levellog.prequestion.domain.PreQuestionRepository;
 import com.woowacourse.levellog.team.application.TeamQueryService;
 import com.woowacourse.levellog.team.application.TeamService;
-import com.woowacourse.levellog.team.domain.ParticipantsFactory;
 import com.woowacourse.levellog.team.domain.Team;
 import com.woowacourse.levellog.team.domain.TeamDetail;
 import com.woowacourse.levellog.team.domain.TeamRepository;
@@ -70,7 +69,7 @@ abstract class ServiceTest {
     protected MemberService memberService;
 
     @Autowired
-    protected OAuthService oAuthService;
+    protected LoginService loginService;
 
     @Autowired
     protected LevellogService levellogService;
